the `priority_queue` is only used by the `executor`.
so rearrange headers accordingly.
